% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{annot_g}
\alias{annot_g}
\title{The result of gff3_preprocessing of gff3 file
A list containing all necessary information from a gff file for
adjusting_HLToInt and visualization.}
\format{
A list with 2 items:
\describe{
\item{data annotation:}{a data frame with 5853 rows and 6 variables
\describe{
\item{region:}{the region from the gff file}
\item{start:}{the start of the annotation}
\item{end:}{the end of the annotation}
\item{strand:}{the strand of the annotation}
\item{gene:}{the annotated gene name}
\item{locus_tag:}{the annotated locus tag}
}
}
\item{genome length:}{a numeric vector containing the length of the genome}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifiComparative}
}
\usage{
data(annot_g)
}
\description{
The result of gff3_preprocessing of gff3 file
A list containing all necessary information from a gff file for
adjusting_HLToInt and visualization.
}
\keyword{datasets}
