% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{df_mean_minimal}
\alias{df_mean_minimal}
\title{The result of adjusting_HLToInt for stats_df_comb_minimal and annotation
example data
A data frame containing the output of adjusting_HLToInt as a data frame}
\format{
A data frame with 52 rows and 15 variables:
\describe{
\item{position:}{The bin/probe specific position}
\item{region:}{the region from the gff file}
\item{gene:}{the annotated gene name}
\item{locus_tag:}{the annotated locus tag}
\item{strand:}{The strand specific}
\item{fragment_HL:}{The half-life fragment the bin belongs}
\item{fragment_int:}{The intensity fragment the bin belongs}
\item{position_frg_int:}{The position of the first fragment and the last
position of the last fragment}
\item{mean_HL_fragment:}{The mean half-life value of the respective
half-life fragments}
\item{mean_int_fragment:}{The mean intensity value of the respective
intensity fragments}
\item{log2FC(decay_rate):}{log2FC(decay(condition1)/decay(condition2))}
\item{Log2FC(HL)-Log2FC(int):}{log2FC(decay_rate/intensity)}
\item{log2FC(synthesis_rate):}{log2FC(decay_rate) +
log2FC(intensity)}
\item{intensity_FC:}{log2FC(mean(intensity(condition1))/mean(
intensity(condition2)))}
\item{p_value:}{indicated by "*" means at least one fragment
either HL fragment or intensity fragment has a significant p_value}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifiComparative}
}
\usage{
data(df_mean_minimal)
}
\description{
The result of adjusting_HLToInt for stats_df_comb_minimal and annotation
example data
A data frame containing the output of adjusting_HLToInt as a data frame
}
\keyword{datasets}
