% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{differential_expression}
\alias{differential_expression}
\title{An example data frame from Synechosystis PCC 6803 differential probes
expression obtained from limma package and only interesting variables were
selected. The data frame was used entirely.}
\format{
A data frame of differential_expression with 55508 rows and 4 variables:
\describe{
\item{position:}{The bin/probe specific position}
\item{strand:}{The strand specific}
\item{logFC_int:}{The bin/probe differential expression}
\item{P.Value:}{The bin/probe p_value adjusted}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifiComparative}
}
\usage{
data(differential_expression)
}
\description{
An example data frame from Synechosystis PCC 6803 differential probes
expression obtained from limma package and only interesting variables were
selected. The data frame was used entirely.
}
\keyword{datasets}
