% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{inp_f}
\alias{inp_f}
\title{The result of loading_fun for stats_se_cdt2 example data
Two data frame containing the output of loading_fun as second element of a
list.}
\format{
A data frame with 500 rows and 49 variables:
\describe{
\item{strand:}{The strand specific}
\item{position:}{The bin/probe specific position}
\item{ID:}{The bin/probe specific ID}
\item{FLT:}{The bin/probe flag for background level}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model is
applied}
\item{flag:}{Information on which fitting model is applied}
\item{position_segment:}{The position based segment}
\item{delay:}{The delay value of the bin/probe}
\item{half_life:}{The half-life of the bin/probe}
\item{TI_termination_factor:}{The termination factor of the bin/probe}
\item{delay_fragment:}{The delay fragment the bin belongs to}
\item{velocity_fragment:}{The velocity value of the respective delay
fragment}
\item{intercept:}{The vintercept of fit through the respective delay
fragment}
\item{slope:}{The slope of the fit through the respective delay fragment}
\item{HL_fragment:}{The half-life fragment the bin belongs to}
\item{HL_mean_fragment:}{The mean half-life value of the respective
half-life fragment}
\item{intensity_fragment:}{The intensity fragment the bin belongs to}
\item{intensity_mean_fragment:}{The mean intensity value of the respective
intensity fragment}
\item{TU:}{The overarching transcription unit}
\item{TI_termination_fragment:}{The TI fragment the bin belongs to}
\item{TI_mean_termination_factor:}{The mean termination factor of the
respective TI fragment}
\item{seg_ID:}{The combined ID of the fragment}
\item{pausing_site:}{presence of pausing site indicated by +/-}
\item{iTSS_I:}{presence of iTSS_I indicated by +/-}
\item{ps_ts_fragment:}{The fragments involved in pausing site or iTSS_I}
\item{event_ps_itss_p_value_Ttest:}{p_value of pausing site or iTSS_I}
\item{p_value_slope:}{p_value of the slope}
\item{delay_frg_slope:}{the slope value of the respective delay fragment}
\item{velocity_ratio:}{Integer, ratio of velocity between 2 delay fragments}
\item{event_duration:}{Integer, the duration between two delay fragments}
\item{event_position:}{Integer, the position middle between 2 fragments
with an event}
\item{FC_HL:}{Integer, the fold change value of 2 HL fragments}
\item{FC_fragment_HL:}{Integer, the fold change value of 2 intensity
fragments}
\item{p_value_HL:}{p_value of the fold change of HL fragments}
\item{FC_intensity:}{Integer, the fold change value of 2 intensity fragments}
\item{FC_fragment_intensity:}{String, fragments involved in fold change
between 2 intensity fragments}
\item{p_value_intensity:}{p_value of the fold change of intensity fragments}
\item{FC_HL_intensity:}{ratio of fold change between 2 half-life fragments
and fold change between 2 intensity fragments}
\item{FC_HL_intensity_fragment:}{fragments involved on ratio of fold
change between 2 half-life fragments and fold change between 2 intensity
fragments}
\item{FC_HL_adapted:}{Integer, the fold change of half-life/ fold change of
intensity, position of the half-life fragment is adapted to intensity
fragment}
\item{synthesis_ratio:}{Integer, the value correspomding to synthesis rate}
\item{synthesis_ratio_event:}{String, the event assigned by synthesis
rate either  Termination or iTSS}
\item{p_value_Manova:}{p_value of the variance between two fold-changes,
HL and intensity}
\item{p_value_TI:}{p_value of TI fragment}
\item{TI_fragments_p_value:}{p_value of 2 TI fragments}
\item{cdt:}{The condition assigned to the experiment here cdt2}
\item{logFC_int:}{The bin/probe log2 fold change of intensity at time 0}
\item{P.Value:}{The bin/probe p_value adjusted }
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifiComparative}
}
\usage{
data(inp_f)
}
\description{
The result of loading_fun for stats_se_cdt2 example data
Two data frame containing the output of loading_fun as second element of a
list.
}
\keyword{datasets}
