% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joining_data_column.r
\name{joining_data_column}
\alias{joining_data_column}
\title{'joining_data_column': joins two data frames from different conditions by
column.}
\usage{
joining_data_column(data)
}
\arguments{
\item{data}{data frame with joined columns from both conditions}
}
\value{
the data frame with joined columns from both conditions with the
corresponding columns: strand, position, ID, intensity.cdt1, position_segment,
half_life.cdt1, TI_termination_factor.cdt1", HL_fragment.cdt1,
intensity_fragment.cdt1, TI_termination_fragment.cdt1, logFC_int, P.Value,
intensity.cdt2, half_life.cdt2, TI_termination_factor.cdt2, HL_fragment.cdt2,
intensity_fragment.cdt2, TI_termination_fragment.cdt2
cdt1: first condition, cdt2: second condition.
}
\description{
'joining_data_column': joins two data frames from different conditions by
column.
}
\examples{
data(data_combined_minimal)
df_comb_minimal <- joining_data_column(data = data_combined_minimal)

}
