% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pen_HL}
\alias{pen_HL}
\title{The result of penalties for df_comb_minimal example data.
A list containing the output from penalties including the logbook and two
penalty objects.}
\format{
A list with 5 items:
\describe{
\item{pen_obj_HL:}{A list with 4 items:
\describe{
\item{logbook:}{The logbook vector containing half-life penalty
information}
\item{HL_penalties:}{a vetor with the half-life penalty and half-life
outlier penalty}
\item{correct:}{a matrix of the correct splits}
\item{wrong:}{a matrix of the incorrect splits}
}
}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(pen_HL)
}
\description{
The result of penalties for df_comb_minimal example data.
A list containing the output from penalties including the logbook and two
penalty objects.
}
\keyword{datasets}
