% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pen_int}
\alias{pen_int}
\title{The result of penalties for df_comb_minimal example data.
A list containing the output from penalties including the logbook and two
penalty objects.}
\format{
A list with 5 items:
\describe{
\item{pen_int:}{A list with 4 items:
\describe{
\item{logbook:}{The logbook vector containing intensity penalty
information}
\item{int_penalties:}{a vector with the intensity penalty and
intensity outlier penalty}
\item{correct:}{a matrix of the correct splits}
\item{wrong:}{a matrix of the incorrect splits}
}
}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(pen_int)
}
\description{
The result of penalties for df_comb_minimal example data.
A list containing the output from penalties including the logbook and two
penalty objects.
}
\keyword{datasets}
