% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifiComparative.r
\docType{package}
\name{rifiComparative}
\alias{rifiComparative}
\title{\subsection{=============================================================================
rifiComparative}{

rifiComparative a successor package of rifi. It compares 2 rifi outputs from
2 different conditions of the same organism.
}}
\description{
rifiComparative was developed to compare 2 rifi outputs from 2 conditions.
The rifi output may differ significantly from 2 conditions. The complexity of
the segments number, position, length and the events make the comparison
between 2 conditions nearly impossible. rifiComparative uses a simple strategy
to generate single segments for both conditions, extract the features and
make them comparable.
}
\details{
Five major steps ate described in rifiComparative:
\enumerate{
\item Joining data
\item penalties
\item fragmentation
\item statistics
\item visualization
}
}
\author{
Loubna Youssar \email{lyoussar@gmail.com}

Jens Georg \email{jens.georg@biologie.uni-freiburg.de}
}
