% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.r
\name{statistics}
\alias{statistics}
\title{statistics check segments significance using statistical test}
\usage{
statistics(data)
}
\arguments{
\item{data}{data frame: data frame output of fragmentation}
}
\value{
A list of two data frames, the first one contains all segments with
p_value and p_value adjusted. The second one removes the duplicated segments
from intensity and could be saved as an excel file.
}
\description{
statistics uses t-test to check HL and intensity segments significance. The
function returns the data frame with p_value p_value adjusted.
The function used is t_test_function.
}
\examples{
data(fragment_int)
stats_df_comb_minimal <- statistics(data= fragment_int)[[1]]
df_comb_uniq_minimal <- statistics(data= fragment_int)[[2]]

}
