% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.r
\name{penalties}
\alias{penalties}
\title{'penalties': find the best set of penalties for half-life and intensity
fragmentation
'penalties' find the best set of penalties for half-life and intensity
fragmentation using dynamic programming. The segmentation of the HL uses the
difference between 2 conditions and the segmentation of the intensity uses
the log2FC. The function uses 4 functions:}
\usage{
penalties(data)
}
\arguments{
\item{data}{data frame with the joined columns from differential
expression and output of rifi_stats}
}
\value{
a list of penalties, the first set for HL and the second for
intensity at time 0.
}
\description{
score_fun_ave.r
make_pen.r
fragment_HL_pen.r
fragment_inty_pen.r
}
\examples{
data(df_comb_minimal) 
pen_HL <- penalties(df_comb_minimal)[[1]]
pen_int <- penalties(df_comb_minimal)[[2]]
}
