% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.r
\name{statistics}
\alias{statistics}
\title{'statistics' uses t-test to check HL and intensity segments significance. The
function return the data frame with p_value p_value adjusted.
The function used is t_test_function.}
\usage{
statistics(data)
}
\arguments{
\item{data}{data frame of the joined column}
}
\value{
a list of two data frames, the first one contains all segments with
p_value and p_value adjusted. The second one removes the duplicated segments
from intensity and could be saved as excel file.
}
\description{
'statistics' uses t-test to check HL and intensity segments significance. The
function return the data frame with p_value p_value adjusted.
The function used is t_test_function.
}
\examples{
data(df_comb_minimal)
df_comb_minimal <- statistics(data= fragment_int)[[1]]
df_comb_uniq_minimal <- statistics(data= fragment_int)[[2]]
write_xlsx(df_comb_uniq_minimal, "df_comb_uniq_minimal.xlsx")
}
